asignaturas salvo el recomendaciones

+++ aplicacionWeb/aplicacionWeb/Model/AsignaturaUsuario/AsignaturaUsuarioGet.cs
@@ -5,7 +5,7 @@ namespace aplicacionWeb.Model.AsignaturaUsuario
     /// <summary>
     /// clase de la asignatura-Usuario
     /// </summary>
-    
+
     public class AsignaturaUsuarioGet
     {
         /// <summary>
@@ -15,17 +15,17 @@ namespace aplicacionWeb.Model.AsignaturaUsuario
         /// <summary>
         /// Nota de la asignatura
         /// </summary>
-     
-        public double Nota { get; set; }
+
+        public decimal Nota { get; set; }
         /// <summary>
         /// horas de estudio
         /// </summary>
-        
+
         public int TiempoEstudio { get; set; }
         /// <summary>
         /// horas de estudio recomendado
         /// </summary>
-      
+
         public int TiempoRecomendado { get; set; }
         /// <summary>
         /// Riesgo en referente a la nota sacada
