+++ tfg_api/Model/AsignaturaUsuario/AsignaturaUsuarioGetSort.cs
@@ -27,7 +27,7 @@ namespace tfg_api.Model.AsignaturaUsuario
         /// Nota de la asignatura
         /// </summary>
 
-        public double Nota { get; set; }
+        public decimal Nota { get; set; }
 
     }
 }
